auth_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
psycopg2-binary
sqlalchemy
prometheus-client